+++ stytra/hardware/video/__init__.py
@@ -62,7 +62,7 @@ class VideoSource(FrameProcess):
 
     """
 
-    def __init__(self, rotation=False, max_mbytes_queue=100, n_consumers=1):
+    def __init__(self, rotation=False, max_mbytes_queue=200, n_consumers=1):
         """ """
         super().__init__(name="camera")
         self.rotation = rotation